pyproject.toml
[project]
name = "thermal-camera-viewer"
version = "0.1.0"
description = "USB thermal camera viewer using OpenCV"
requires-python = ">=3.8"
dependencies = [
    "torch<2.3.0",
    "torchvision<0.18.0",
    "ultralytics",
    "opencv-python",
    "numpy<2",
    "pyusb",
    "pyserial",
]

[tool.setuptools]
py-modules = ["main", "uvc_controls"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
